Cache doppler history for faster server page load.
